+++ harvest_dwer/DWER.sh
@@ -70,12 +70,13 @@ make_dt () {
 }
 
 #------------------------------------------------------------------------------#
-curl --user ${USERNAME}:${PASSWORD} --list-only "${URL}/" -s -o $TMPLST
+curl --user ${USERNAME}:${PASSWORD} --list-only "${URL}/" -s -o ${TMPLST}
 
-mkdir -p ${DATADIR} > /dev/null 2>&1
+if [ -f ${TMPLST} ] ; then
+  mkdir -p ${DATADIR} > /dev/null 2>&1
 
-if [ "$SITENAME" = "flow" ] ; then
-    grep "${YEAR}-${MONTH}-${DAY}" $TMPLST | while read LINE ; do
+  if [ "$SITENAME" = "flow" ] ; then
+    grep "${YEAR}-${MONTH}-${DAY}" ${TMPLST} | while read LINE ; do
         FILE=`echo $LINE | tr -d '\r'`
         DIRN=`echo $FILE | cut -f1 -d~`
         DIRN="${DIRN}-"`echo $FILE | cut -f2 -d~`
@@ -87,54 +88,62 @@ if [ "$SITENAME" = "flow" ] ; then
         curl --user ${USERNAME}:${PASSWORD} "${URL}/${FILE}" -s -o ${DATADIR}/${DIRN}/${TODAY}.csv
     done
 
-    /bin/rm $TMPLST
+    /bin/rm ${TMPLST}
 
     . ./common/finish.sh
     exit 0
-fi
+  fi
 
 #------------------------------------------------------------------------------#
 
-SEARCH="${DAY}/${MONTH}/${YEAR}"
+  SEARCH="${DAY}/${MONTH}/${YEAR}"
 
-cat $TMPLST | while read LINE ; do
+  cat ${TMPLST} | while read LINE ; do
     FILE=`echo $LINE | tr -d '\r'`
-#   echo Fetching ${FILE}
     curl --user ${USERNAME}:${PASSWORD} "${URL}/${FILE}" -s -o ${DATADIR}/${FILE}
-    DIRN=`echo ${FILE} | sed -e 's/\.csv$//'`
-    if [ "${DIRN}" != "${FILE}" ] ; then
-      mkdir -p ${DATADIR}/${DIRN} > /dev/null 2>&1
-      if [ -f ${DATADIR}/${DIRN}/${TODAY}.csv ] ; then
-         TCL=`tail -1 ${DATADIR}/${DIRN}/${TODAY}.csv | cut -f1 -d,`
-         export CURLAST=`makeiso2 "$TCL"`
+    if [ -f ${DATADIR}/${FILE} ] ; then
+      DIRN=`echo ${FILE} | sed -e 's/\.csv$//'`
+      if [ "${DIRN}" != "${FILE}" ] ; then
+        mkdir -p ${DATADIR}/${DIRN} > /dev/null 2>&1
+        if [ -f ${DATADIR}/${DIRN}/${TODAY}.csv ] ; then
+          TCL=`tail -1 ${DATADIR}/${DIRN}/${TODAY}.csv | cut -f1 -d,`
+          export CURLAST=`makeiso2 "$TCL"`
+        else
+          export CURLAST=000000000000
+        fi
+
+        echo -n > ${DATADIR}/${DIRN}/${TODAY}.csv
+        grep "${SEARCH}" ${DATADIR}/${FILE} | while read L2 ; do
+           INDT=`echo $L2 | cut -f1 -d,`
+           REST=`echo $L2 | cut -f2- -d,`
+           ISODT=`makeiso "$INDT"`
+           NEWDT=`make_dt "$INDT"`
+
+           if [ $ISODT -gt $CURLAST ] ; then
+             echo $NEWDT,$REST >> ${DATADIR}/${DIRN}/${TODAY}.csv
+             set_data_date "$NEWDT"
+             log_last_update
+
+             CURLAST=$ISODT
+           fi
+        done
+
+        count=`wc -l ${DATADIR}/${DIRN}/${TODAY}.csv | cut -f1 -d\ `
+        if [ $count -eq 0 ] ; then
+           /bin/rm ${DATADIR}/${DIRN}/${TODAY}.csv
+           /bin/rmdir ${DATADIR}/${DIRN} > /dev/null 2>&1
+        fi
       else
-         export CURLAST=000000000000
+        echo unexpected file \"${FILE}\"
       fi
 
-      echo -n > ${DATADIR}/${DIRN}/${TODAY}.csv
-      grep "${SEARCH}" ${DATADIR}/${FILE} | while read L2 ; do
-         INDT=`echo $L2 | cut -f1 -d,`
-         REST=`echo $L2 | cut -f2- -d,`
-         ISODT=`makeiso "$INDT"`
-         NEWDT=`make_dt "$INDT"`
-         if [ $ISODT -gt $CURLAST ] ; then
-           echo $NEWDT,$REST >> ${DATADIR}/${DIRN}/${TODAY}.csv
-           CURLAST=$ISODT
-         fi
-      done
-      count=`wc -l ${DATADIR}/${DIRN}/${TODAY}.csv | cut -f1 -d\ `
-      if [ $count -eq 0 ] ; then
-         /bin/rm ${DATADIR}/${DIRN}/${TODAY}.csv
-         /bin/rmdir ${DATADIR}/${DIRN} > /dev/null 2>&1
-      fi
-    else
-      echo unexpected file \"${FILE}\"
+      /bin/rm ${DATADIR}/${FILE}
     fi
-    /bin/rm ${DATADIR}/${FILE}
-done
+  done
 
-/bin/rm $TMPLST
+  /bin/rm ${TMPLST}
 #------------------------------------------------------------------------------#
+fi
 
 . ./common/finish.sh
